Start end walk with FK Relationship to Walker

+++ Services/Walks/WalksService.cs
@@ -1,9 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Data;
 using Data.Entities;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using Models.Walks;
 
@@ -12,10 +14,18 @@ namespace Services.Walks
     public class WalksService : IWalksService
     {
         private readonly ApplicationDbContext _db;
+        private readonly int _userId;
 
-        public WalksService(ApplicationDbContext db)
+        public WalksService(IHttpContextAccessor httpContext,ApplicationDbContext db)
         {
             _db = db;
+             var userClaims = httpContext.HttpContext.User.Identity as ClaimsIdentity;
+            var value = userClaims?.FindFirst("Id")?.Value;
+            var validId = int.TryParse(value, out _userId);
+            if (!validId)
+            {
+                throw new Exception("Attempted to build RatingService without User Id Claim");
+            }
         }
 
         public async Task<bool> NewWalkAsync(CreateWalks model)
@@ -26,9 +36,9 @@ namespace Services.Walks
                 DistanceWalked = model.DistanceWalked,
                 Lattitude = model.Lattitude,
                 Longitude = model.Longitude,
-                WalkerName = model.WalkerName,
+                WalkerId = model.WalkerId,
                 OutsideTemp = model.OutsideTemp,
-                WalkStarted = model.WalkStarted,
+                WalkStarted = DateTime.UnixEpoch,
                 WalkEnded = model.WalkEnded
             };
             _db.Walking.Add(entity);
@@ -45,7 +55,7 @@ namespace Services.Walks
                 DistanceWalked = c.DistanceWalked,
                 Lattitude = c.Lattitude,
                 Longitude = c.Longitude,
-                WalkerName = c.WalkerName,
+                WalkerId = c.WalkerId,
                 OutsideTemp = c.OutsideTemp,
                 WalkStarted = c.WalkStarted,
                 WalkEnded = c.WalkEnded,
@@ -57,12 +67,12 @@ namespace Services.Walks
         public async Task<bool> UpdateWalkAsync(WalksUpdate req)
             {
                 var entity = await _db.Walking.FindAsync(req.Id);
-                entity.Id = req.Id;
+                if(entity is null) return false;
                 entity.DogId = req.DogId;
                 entity.DistanceWalked = req.DistanceWalked;
                 entity.Lattitude = req.Lattitude;
                 entity.Longitude = req.Longitude;
-                entity.WalkerName = req.WalkerName;
+                entity.WalkerId = req.WalkerId;
                 entity.OutsideTemp = req.OutsideTemp;
                 entity.WalkStarted = req.WalkStarted;
                 entity.WalkEnded = req.WalkEnded;
@@ -73,9 +83,63 @@ namespace Services.Walks
         public async Task<bool> DeleteWalkByIdAsync(int Id)
         {
             var walks = await _db.Walking.FindAsync(Id);
+            if(walks is null) return false;
             _db.Walking.Remove(walks);
             var changed = await _db.SaveChangesAsync();
             return changed == 1;
         }
+
+        public async Task<IEnumerable<WalksDetail>> GetWalksByCurrentIdAsync()
+        {
+            return await _db.Walking.Include(w => w.Dog).Where(c => c.WalkerId == _userId).Where(c=>c.WalkStarted == DateTime.UnixEpoch).Select(c => new WalksDetail
+            {
+                DogId = c.DogId,
+                DistanceWalked = c.DistanceWalked,
+                Lattitude = c.Lattitude,
+                Longitude = c.Longitude,
+                WalkerId = c.WalkerId,
+                OutsideTemp = c.OutsideTemp,
+                WalkStarted = c.WalkStarted,
+                WalkEnded = c.WalkEnded,
+                DogName = c.Dog.Name
+
+            }).ToListAsync();
+        }
+
+        public async Task<IEnumerable<WalksDetail>> GetOngoingWalksByCurrentIdAsync()
+        {
+            return await _db.Walking.Include(w => w.Dog).Where(c => c.WalkerId == _userId).Where(c=>c.WalkStarted != DateTime.UnixEpoch && c.WalkEnded == null).Select(c => new WalksDetail
+            {
+                DogId = c.DogId,
+                DistanceWalked = c.DistanceWalked,
+                Lattitude = c.Lattitude,
+                Longitude = c.Longitude,
+                WalkerId = c.WalkerId,
+                OutsideTemp = c.OutsideTemp,
+                WalkStarted = c.WalkStarted,
+                WalkEnded = c.WalkEnded,
+                DogName = c.Dog.Name
+
+            }).ToListAsync();
+        }
+
+
+        
+public async Task<bool> EndWalkByIdAsync(int id)
+        {
+            var entity = await _db.Walking.FindAsync(id);
+                if(entity is null) return false;
+            entity.WalkEnded = DateTime.Now;
+             return await _db.SaveChangesAsync() == 1;
+        } 
+        public async Task<bool> StartWalkByIdAsync(int id)
+        {
+            var entity = await _db.Walking.FindAsync(id);
+                if(entity is null) return false;
+            entity.WalkStarted = DateTime.Now;
+            entity.DistanceWalked =0;
+             return await _db.SaveChangesAsync() == 1;
+        } 
+        
     }
 }
